backend/requirements.txt
Django>=5.0
djangorestframework>=3.15.0
django-cors-headers>=4.3.1
pandas>=2.2.0
openpyxl>=3.1.2
python-dotenv>=1.0.1
google-genai>=0.2.0
pydantic>=2.8,<3
json-repair
dateparser>=1.2
python-dateutil>=2.8.2
numpy>=1.24.0